fix
文字が指定領域からはみ出す場合に，収まる部分も描画されない問題を修正

+++ DirectXFont.hpp
@@ -154,7 +154,7 @@ namespace dx9 {
 		// 文字列strのoffset番目の文字から，長さlimitに1行で入る文字数を取得
 		// 1行の長さは，lengthに書き込まれる
 		// limit < 0.0fの場合，長さの指定を無視する
-		int GetStrLength(const wchar_t* str, size_t offset, float limit, size_t &length);
+		int GetStrLength(const wchar_t* str, size_t offset, float limit, int &length);
 
 
 	};
